refactor: modularize utilities into dedicated packages for response and subscription handling
